show the modification time of each page individually
The old way of setting last_updated for the output showed a single
value on all pages, which makes it more difficult to know how up to
date a given resolution or policy is. The output was also formatted as
binary strings, which rendered poorly with b'' wrapped around the
actual timestamp.

This change uses the git history of each page to show when it was
modified. For pages generated from the project list, show the
modification time of the input data file. For pages rendered from
templates, without a source file, use the current build time. Always
return the value as a unicode string.

Change-Id: I82849176775484328b3939b4e0c66aaf7a35f49d
Signed-off-by: Doug Hellmann <doug@doughellmann.com>