Explicitly declare Stein supported runtimes
This adds a PTI section to have an official declaration of
the targeted runtimes for the release. This should also be retroactively
added to other cycle sections so we have a historical record of what was
supported for that release.

Javascript section omitted because I have zero experience with Node and
NPM, but it would be good to expand on this to include all of our
official runtimes.

Change-Id: I11b136a9781dd1a77df536f099df10a6dd524e06
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>