Indicate relmgt style for each deliverable
The Release Management team is handling release management for
deliverables from all official OpenStack teams. However, there
are a number of exceptions:

- deliverables that are not using tags or branches for their
  'release', or that are directly published (docs, specs,
  cookiecutters...)

- deployment deliverables that are released on a specific
  marketplace (Chef supermarket, Charm store...) and that are
  not relying on the OpenStack release management team help

To facilitate tracking the intention of the teams for their
deliverables, this change introduces a "release-management:"
key to the deliverable entries. By default (if the key is not
present), the deliverable would be handled by the OpenStack
Release Management team using the openstack/release repository.

But the key could also contain:

- release-management: none (the deliverable does not need any
  release management, it does not use tags or series branches)

- release-management: external (the deliverable is published
  on a separate publication platform, and its release is managed
  by its project team directly; this should be limited to corner
  cases like deployment tools)

This change applies the key to already-known exceptions.

Change-Id: I8ffe1de25cc3252a2d9e6acb21f24934c313cfcb
Task: 22262