Reword open source definition in the Four Opens
The current phrasing states "Truly open source software is not feature
or performance limited and is not crippled."

Crippled in that context means it is not disabled or limited in any
manner by the fact the software is open source. As a non-native English
speaker, the first impression on looking up that adjective on the
internet is that it is an English word used to describe people with
disabilities, sometimes in a pejorative manner. Its translation to the
Portuguese word 'aleijado' which does not seem respectful, translating
back to 'crippled', 'lame' or 'gammy'.

As it led me to confusion, it may lead others too, specially in a
community where almost every country in this planet is represented.

This patch suggests the removal of that part of the sentence, leaving
"Truly open source software is not feature or performance limited."
which still has the same meaning that is clear and thus does not need
to be emphasized by that adjective.

Change-Id: I29b6cea609ae9db9c9ff678c28744de4570486ca