Clarify scope of new projects
Now that we have the OpenStack project map, it is clear that our
description of the required scope for new projects really only
encompasses the main 'OpenStack' bucket on the map, when we have in
practice included several other classes of projects and always intended
to continue doing so.

https://www.openstack.org/assets/software/projectmap/openstack-map.pdf

This change documents the allowed purposes of new projects, using the
categories defined in the project map to describe them.

Change-Id: I02e91254fea83acbdb694dffefc7d864f0bad105