update home page for openstackclient
The OSC team does not use the wiki page currently listed as the home
page, and it is clearly marked deprecated. Linking to it here is causing
confusion for users looking for the most recent information about bug
trackers and other details. This change updates the link to point to the
documentation for the project instead.

Change-Id: Ie3543c3d045d146243e2d8fc1ae352824504262c
Signed-off-by: Doug Hellmann <doug@doughellmann.com>