Update charter to include PTL appointment
We've had a couple projects go through changes in leadership, for
which we do have a couple formalities, but they weren't written down.
This commit updates the charter to include the items that need to be
taken care of if a PTL steps down and appoints another individual to
the role.

Change-Id: Ie162d3c66abc698bf31e7224530869a836e111e9