Clarify wording of Python update resolution
The previous wording introduced unneccesary confusion about when the
goal would be set. Clarify that the goal is set at the beginning of a
release cycle as usual, and should be completed by the end of the
release cycle.

Also, avoid using the word 'support' to refer to which versions of
Python we run unit tests on.

Change-Id: I461ba0c0d4bc1e3d7e2b71ab614003733afdd4ea