No release management for deprecated repositories
Some deprecated deliverables have no release management anymore,
but have not been removed from governance yet (as they are still
shipped in some maintained branches).

Neither release-management:none nor release-management:external
accurately describe that situation, so we introduce a new
release-management:deprecated to account for those deliverables
that do not need release management because they are on the
deprecation track.

Examples of this are tempest-lib, devstack-plugin-zmq and
devstack-plugin-pika.

Change-Id: I9c0f4270835d0df127ac2021a10db0f8660aecab