Cloud Vision: clarify meaning of design goals
Feedback on this section from a couple of different sources reveals that
the current description fails to make clear that the design goals
represent capabilities that we want OpenStack to present to
applications, and aren't necessarily descriptive of the OpenStack
service implementations themselves. Tweak the wording to help readers
interpret the goals in the correct context.

Change-Id: I5e2fe12dca016f0e9039cdf4290060ed51bc10ce