Amend reference/PTI to add openSUSE
When the Ubuntu and CentOS were added in the PTI as "supported
distros", SUSE was already one of the most-popular distros [1],
and therefore deserves listing here.

I don't believe the PTI should refer to "popular" term for distros,
as this isn't a popularity contest. Instead, we should allow any
distribution where community effort is provided.

In this case, openSUSE images are present in infra, and SUSE contributors
are using devstack on openSUSE. Next to that, deployment projects like
openstack-ansible or openstack-helm are making use of openSUSE.

Last, but not least, some projects already take into account openSUSE
in their regular work (for example, when the nova team contacts the
distributions at the beginning of the cycle to know their latest
supported libvirt [2]).

I believe all these arguments are enough to warrant the addition of
openSUSE in the list of distributions.

[1]: SUSE was ranked 4th in that survey (with SLES), after RHEL (3rd).
[2]: http://lists.openstack.org/pipermail/openstack-dev/2018-September/135007.html

Change-Id: I1ac3b7477b5d4cb4af75c319b4232d20b4b4b284