apply 7 day rule to formal-vote items
Ensure that the charter rule that formal-vote items stay open at least
7 days is applied. This change interprets that rule to mean that if
the majority votes in favor of a change within the 7 day period, the 3
day waiting period will include any of the remaining 7 days. So if the
majority is reached on day 1, the patch must stay open 7 days and if
the majority is reached on day 7 the patch must stay open an
additional 3 days.

Change-Id: I48841570ff19b779b2e35de8218a8f56b408df92
Signed-off-by: Doug Hellmann <doug@doughellmann.com>