Add link to activity metrics to team pages
We are looking at removing team details from the openstack.org/software
pages, and linking to the team pages on governance.o.o/tc instead.

Add a link to team activity metrics (as displayed by Stackalytics)
to the team pages, so that people can still find that data if they care
about it.

Change-Id: I8fa5cf7e77174b9ddab2e184e839159afcec2ad4