Describe the business value of using unified limits
Recent threads on the openstack-discuss mailing list [0] suggest
revisting the idea of formatting the help-wanted list to represent job
requests that describe the value of an initiative to potential
sponsors. This was feedback the OpenStack Technical Committee received
from the Board of Directors during the Vancouver summit in 2018.

This is an attempt at doing that with an existing cross-project
initiative to consume unified limits.

[0] http://lists.openstack.org/pipermail/openstack-discuss/2019-January/002252.html

Change-Id: I0cb4a0b1c0366afaa6d451d6c498d5b45b1fe242