Add Mohammed Naser nomination as chair
I would like to nominate myself to serve as the chair of the TC
for the upcoming session.

First of all, I'd like to thank Doug for his service as the chair
for the past few periods of time.  He has been instrumental in
helping our TC, provided a lot of leadership and guidance and
he was very instrumental in helping me understand this role by
helping him as vice-chair.

We have a lot to work on as a team and I would like to explore
more and new options to communicate more often and follow-up on
our projects within the TC.  As time goes by, the role of the TC
to validate new projects seems to have decreased, and it seems
that the community would like a more "hands-on" involvement with
actionable 'artifacts' that come out of the TC.

That's not to say that we haven't built any of that over the past
few months, we've come up with a lot of very important documents
to help define a lot of things, but I'd like us to focus on working
with our teams to help push OpenStack with our community in the
best and most succcesful path possible.

We have a community of long term contirbutors that we trust to know
the best decision to take, however, there is still a lot of doubt
from our new-comers and perhaps a lot of questions on "how should
we deal with this?" -- as a technical committee, we need to be there
for our developers to help steer them in the best direction.

With the changes within the OSF, it's also more important than ever
to be aware of what is happening on the foundation level on how it
affects us, as well as working with other foundation-level projects
to find value that benefits all of us (which we already do, in the
case of Zuul).

Thank you for your consideration.

Change-Id: I92c867c7237295bc6fea731b656d5d2c69e01926