Elaborate on the business value of goal champions
Recent threads on the openstack-discuss mailing list [0] suggest
revisiting the idea of formatting the help-wanted list to represent
the value each initiative provides a sponsor. This feedback was given
to the Technical Committee during the Vancouver summit in 2018.

This is an attempt at describing the value sponsoring a goal champion
has on a business.

This commit also amends the document introduction to include intended
audiences and which parts of the document they might find most useful.

[0] http://lists.openstack.org/pipermail/openstack-discuss/2019-January/002252.html

Change-Id: Idfa593a12550f8cdc3a5f617b8271a30e1d7a3b9