Elaborate on the business value of Glance
Recent threads on the openstack-discuss mailing list [0] suggest
revisiting the idea of formatting the help-wanted list to represent
the value each initiative provides a sponsor. This feedback was given
to the Technical Committee during the Vancouver summit in 2018.

This is an attempt at describing the value contributing to Glance
has on a business.

[0] http://lists.openstack.org/pipermail/openstack-discuss/2019-January/002252.html

Change-Id: I49d36758c3a7a5044f2a46a5a869819e36c24cca