Elaborate on the business value of Designate
Recent threads on the openstack-discuss mailing list [0] suggest
revisiting the idea of formatting the help-wanted list to represent
the value each initiative provides a sponsor. This feedback was given
to the Technical Committee during the Vancouver summit in 2018.

This is an attempt at describing the value contributing to Designate
has on a business.

[0] http://lists.openstack.org/pipermail/openstack-discuss/2019-January/002252.html

Change-Id: Ida1f2334db892107051d093d013e937cdee5f709