Elaborate on the business value of infra admins
Recent threads on the openstack-discuss mailing list [0] suggest
revisiting the idea of formatting the help-wanted list to represent
the value each initiative provides a sponsor. This feedback was given
to the Technical Committee during the Vancouver summit in 2018.

This is an attempt at describing the value improving the OpenStack CI
system has on a business.

[0] http://lists.openstack.org/pipermail/openstack-discuss/2019-January/002252.html

Change-Id: Icb0560e74afe4a62193aed573634280be645524a