Explicitly declare Train supported runtimes.
The TC added openSUSE Leap as a PTI OS in the Stein cycle, and at a
TC ad-hoc meeting [1] the TC defined python 3.6 and 3.7 as the required
versions for the Train cycle.

Golang is still in a rapid development state, and as a result we do not have
a recommended stable golang version for all projects.

1 - http://eavesdrop.openstack.org/meetings/tc_python3/2019/tc_python3.2019-03-07-21.00.html

Change-Id: Ibf0c77f409fb83ac20ddaac4a96f21f7d6516165
Signed-off-by: Graham Hayes <gr@ham.ie>