Appoint Divya K Konoor as PowerVMStackers PTL
PowerVMStackers did not have any PTL candidate in the required candidacy
timeframe and as a result has triggered the leaderless team process.

Divya K Konoor did submit a candidacy[1] but too late, and she only had
commits too recent to be a valid candidate. The PTL for the Stein cycle
confirmed[2] that she would make a great PTL.

[1] https://review.openstack.org/#/c/643003
[2] http://lists.openstack.org/pipermail/openstack-discuss/2019-March/003770.html

Change-Id: I9f3c7a38e3eb83f9a67169204d71a206d2f3d132