members: add Alexandra Settle as vice-chair
Alexandra has been a member of our community since the Icehouse
release cycle, involved in the documentation project and has
recently joined the technical committee.

The exposure working with the OpenStack documentation team means
she has a lot of user facing experience and brings a very important
point-of-view within the technical committee.

She has agreed to serve as vice-chair for the upcoming session,
so I would like to push up this patch to get that documented.

Change-Id: I7e77191da1af1507cacae653747d4b6d568e1098