Remove stable:follows-policy tag for Trove
The reason for this proposal is that Trove team would like to manage the
way how stable branch is maintained, the original discussion is here[1].
There are several things argued in the discussion, self-approve policy
and stable branch permission. Removing the stable:follows-policy tag is
just aiming at solving the latter.

I don't know the history that when Trove added the tag, but I personally
don't think that applies to Trove for now. The tag is useful especially
for the projects that are adopted universally in production, which means
there are lots of people care about the stability and maintainability of
the projects. It's good for those projects to follow some policies to
achieve that.

However, Trove is no longer a project that was backed by Tesora, the
previous PTL came from Tesora left, all of the core reviewers left.
Until now, I've never heard there is company using Trove in production
except a private cloud in Samsung, maybe there are some others, but
seems they don't care about the open source community[2]. As mentioned
in my PTL candidacy, the highest priority for Trove is to give the
potential users confidence that the project is still actively
maintained, the team(well, mostly myself) is trying very hard to improve
DevStack and improve the way to build Trove guest image, in the near
future, I will solve the security issue as well, because those issues
were always asked when the developers or operators started the project
evaluation. With all that said, we need to decide what is going to be
backported by ourselves, sometimes even features or patches that don't
satisfy the stable branch policy. I don't see there are many advantages
to involve Stable Branch Maintenance team.

Hopefully, the project could get the tag back in the future.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2019-April/004991.html
[2] http://lists.openstack.org/pipermail/openstack-discuss/2019-January/001934.html

Change-Id: I14e1377e87b3fa16c1414ae9a780144d48ca213e