Add Upstream Institute training image repo under Docs
Adding the repository containing the files for the VM image we use for the
OpenStack Upstream Institute training. As the OpenDev changes are taking place
soon the intent is to have the training relevant repositories under one umbrella
to avoid future confusions on where the resources are and where they belong.

Change-Id: I6271a3034ba00e37f90b344af814ecdeac196b05