Update the 'stable:follows-policy' tag doc for re-obtaining process
We have 'stable:follows-policy' tag to mark the projects to
follow the stable branch policies. We have defined the tag
application and removal process but process to re-obtaining
the tag is not defined.

This commit updates the doc with more clarification about re-obtaining
the tag and when project remove the tag what that actually means.

Change-Id: I4b4a87f7bb3fc635dd6b27310e1401d06f51b962