Finalise the transition of the docs team to a SIG
The documentation team is officially being disbanded as
an official project team in favor of a SIG. A "SIG" is a group
of people with a shared, special interest, pushing common
objectives. This review moves the remaining docs to deliverables
of the newly formed Technical Writing SIG.

The docs team has been a group of technical contributors
maintaining guidelines, tooling, and best practices for
the project teams whilst the project teams have been responsible
for producing documentation for each component since the Pike
release.

Whilst documentation is still considered imperative to an OpenStack
release, the docs team is no longer in charge of producing the
product documentation and therefore should be disbanded in favor
of a SIG to reflect their current working model in an official
capacity.

Depends-On: https://review.opendev.org/#/c/691277/
Change-Id: I1c47c7d09b0ef89b7d737b0a29b828fbbf9e141a
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Signed-off-by: Alexandra Settle <asettle@suse.com>