Rewrite IPv6 goal to be clearer that we already v6
The original IPv6 community goal is laudable, but misleading. It
behooves us to avoid incorrectly implying that we lack solid IPv6
support. We've provided solutions for IPv6 connectivity to user
resources for a very long time and even have users successfully
deploying OpenStack environments with IPv6 control planes. What we
can't demonstrate with confidence yet is that OpenStack is usable in
IPv6-only environments (those with no traditional IPv4 addressing),
so we should clarify that the goal is about taking that next step.

Change-Id: Ic657850a06ec70dedf153ac754cfa707dbee2b90