Use opendev for repository links
Update the :repo: shorthand role to use opendev instead of cgit links
in project team pages.

Make sure that the other content under the reference/ directory is
updated to use :repo: instead of cgit links (except the one in
pti/javascript.rst which should really point to the NPM package).

Change-Id: I1009ba2a0f8d0fd1dab26f05241903b71c63e4fb