Nest tag project output to make sorting look sane
The list of projects was sorted by team name, but only the project names
were included in the output, causing the list to seem oddly formatted.
This adds the team name as its own top level bullet item to each so it is
clear on ownership and looks more logically sorted in the list.

Change-Id: I0b892379299709fcefae668f811e746a007354e6
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>