Defining popup teams
Following the discussion at the Forum in Denver and the TC meeting
at the PTG (also in Denver) one of the outcomes[1] of those discussions
was to propose a lightweight framework for popup teams.

To avoid repeating past failures in cross-project work, this proposal
pushes forward a relatively low barrier to entry, with the TC making an
early decision on whether the team objective sounds like a good idea
for OpenStack, rather than approving detailed implementation specs. It
focuses on supporting the popup teams rather than "blessing" them.

[1] http://lists.openstack.org/pipermail/openstack-discuss/2019-May/006225.html

Change-Id: I39a4c3814ec3442c60c91055b41b0876976aaeab