Adding Image Encryption as a popup team
The Image Encryption popup team is what pushed us to define
a framework around the concept of temporary teams that need
some support from the TC to more easily achieve its goals.

Remove requirement to list all affected project teams, as this
should really be done as part of the popup team work.

Change-Id: Ibf164176a4a07b41ab9840e2d20d8f3bc76ca893