Add 'U' Release name poll
In order to accommodate the challenges of naming a 'U' release in China
some specific clarifications for acceptable names and geographic region
have been defined.  I believe these are well within the spirit of the
existing criteria.

Specifically:

* Spelling. Currently, the most popular system of romanization of Mandarin
  is the Hanyu Pinyin. Unfortunately, Hanyu Pinyin does not produce any
  romanization which begins with the letter U. In order to allow for U
  names, romanizations produced by other recognized systems shall be
  admissible. For the sake of precision, the accepted systems shall be
  Hanyu Pinyin [PY] (official system of China since 1958), Gwoyeu Romatzyh
  [GR] (official system of China 1929-1958), Wade–Giles romanization, and
  Yale romanization. In practice, only the Gwoyeu Romatzyh system produces
  words starting with the letter U. Reference:
    http://pinyin.info/romanization/compare/gwoyeu_romatzyh.html
* In addition to the romanized names produced by one of the recognized
  romanization systems, traditional and customary spellings of names are also admissible.

We've also defined the geographic region as:
1 The municipality of Shanghai. (Note that Shanghai, being a top level
  administrative division of China, is not part of a province. See
  https://en.wikipedia.org/wiki/Direct-administered_municipalities_of_China#Current_PRC_municipalities
2 All top level administrative divisions of China which borders Shanghai.
  (Jiangsu, Zhejiang)
3 All All top level administrative divisions of China which borders those in
  2 above. (Shandong, Anhui, Jiangxi, Fujian)

Eric Kao has kindly volunteered to provide cultural and language
knowledge.  Thanks Eric.

The timing of this poll is a little different in that I've allowed 1
week between the nominations period closing and the opening of voting.
This is to ensure that there is adequate time to translate the poll and
ensure that all candidate names are correctly classified.

Change-Id: I3f7f64a2a2a6c244fbecb381228c1c1827d79673