Only get git timestamp for generated files
openstackdocstheme now gets the last modified timestamp for files by
default. This makes most of our sphinx extension for this redundant. We
do have a few generated files that we perform special handling that
openstackdocstheme does not know about, so this updates our extension to
optimize for only checking the files that will not be covered by the
theme.

This also adds a bindep file to fix a doc build issue with a font
missing. Gate jobs used to use a common bindep-failback.txt file if
there wasn't a local one present. That mechanism was recently removed,
and it appears now the docs build will not pass unless a needed fonts
package is installed that is not there by default.

Change-Id: I454f4d92e397d6410cb869498f348b86ae26ca0a
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>