Fix follow up comment of pdf goal
doc is per release version things and sometime
those are updated with the backport changes to
stable branches.

For example nova rocky 18.0.0 and 18.2.0 can have
changes in doc, so pdf for both release should
contain the release version explicitly.

This is fixing the comment given in original goal
proposal patch
- https://review.opendev.org/#/c/647712/8/goals/train/pdf-doc-generation.rst@55

Change-Id: I9c13b89574c1ae9855b49bc9a45ab3756b8b1961