Add empty deliverables list to (retired) refstack
When the Refstack team was retired, its only deliverable repos were
adopted by the Interop working group of the OSF board of directors
and so no deliverables were added to the legacy.yaml set. Insert an
empty deliverables list to be explicit about it, since election
tooling expects every team in this file to have a list of
deliverables and altering the tooling to ignore a missing
deliverables key makes us fragile in the face of errors of omission.
Explicit rather than implicit.

Change-Id: I6f8fc02cb1d280749691c28d8004af9b628f9ed3