Add URL to rpm-packaging
Rpm-packaging is the only project to not have a url.

Without this patch, it will be impossible to enforce a schema
validation on the presence of an URL.

This is a problem as I believe that we should validate the
presence of an URL, because a link is one of the best ways to
get started with a problem.

This fixes it by providing the rpm-packaging landing page,
pointing to each of the project's source code and documentation.

Change-Id: I1cb1d9b66a8c7278ff481e6c25239ea8fd339eab