No release in train for abandoned deliverables
Some deliverables have been abandoned and will not see a
release in Train. Record that in the repository so that
the release team knows to ignore them.

instack-undercloud: EmilienM said "this one has been retired
in Stein, no release in Train."

tripleo-common-tempest-plugin: EmilienM said "this one is
abandoned, no release"

Change-Id: I21a584a15c129ab5c3ffb6c32d77bdac7076422b