Add Swift mission into governance
Swift is one of the few projects to not have a mission.

Without this patch, it will be impossible to enforce a schema
validation on the presence of a mission.

This is a problem as we could use the mission field for
some different purposes, like o.o website.

This fixes it by providing some content extracted from the
wiki of Swift.

Change-Id: I8f64b16248b3ea4dbf7a376a90a20a735b745ab3