linters: enforce ATC membership after January 2020
In the past, we did not have any enforcement for extra ATCs in
terms of foundation membership.  This has resulted in entries
being added which are not foundation members that cannot be
ATCs.

This patch adds a check to ignore our old records (until they
expire out in January 2020) and enforce it on any newly added
ATCs.

Change-Id: Id1f6d5b0bffb41c952e140372a3594669359fbdb