Align release naming process with practice
The actual conduct of the release naming process has diverged from that
which is documented. Bring them back into alignment by giving the naming
co-ordinator and the TC an explicit role in removing names that aren't
in line with the Code of Conduct's injunction to be considerate of our
global audience, prior to their inclusion in the poll.

Change-Id: I4a4b17bbbe2e3168c70b67594639946895ff4d1f