Clarify proscription of generic release names
The fourth release name criterion arguably implies that generic names of
types of geographic features are not eligible, but is ambiguous and has
been the source of confusion. Reduce this by stating it explicitly.

Arguably this rule would have required an exception to be granted for
Pike, and thus could be considered a change in policy (though
interpretations on this seem to differ based on whether you are from
Boston or not).

Change-Id: I158a32b76aee559c194a4bcafa1482e0426e87f3