Any member of the TC can approve fast tracked patches
As elected representitves, we should trust in the TC members to do what is best
for the OpenStack project. For the fast tracked changes, we should not require
the chair of the TC to approve typos, or reverts, and instead trust the TC
to only use their +W where it is appropriate.

This would cause a change to the gerrit ACLs to add +W to members
of the "tech-committee" group on review.opendev.org.

Change-Id: I010318bf872034fe6a550f1d7d1050e7e95bc76e
Signed-off-by: Graham Hayes <gr@ham.ie>