Update vulnerability:managed policy
The vulnerability:managed tag's requirement #5 for a threat analysis
was untenable, and too significant a barrier for most teams to
overcome. Even most deliverables already covered by the tag never
had one performed in the ~4 years since it was added. The makeup of
our community was also somewhat different at the time it was
written: we no longer have a group of volunteers like the OSSG with
bandwidth to perform and review threat analyses (arguably we never
did), but also the VMT is well-staffed and the volume of
vulnerability reports it deals with has trailed off significantly.
In 2012 we published 20 advisories, and this climbed to 41 issued
within 2014. In contrast, we published 2 advisories in 2018, and 6
so far in 2019. The VMT has also regularly consulted with teams on
suspected vulnerabilities (on a best-effort basis) even on
deliverables lacking the vulnerability:managed tag. As a result,
while the softening of #5 from a hard requirement to a
recommendation could result in a significant increase in the number
of covered deliverables, the increase in caseload for the VMT
shouldn't be overwhelming.

Additionally, so as to offset some of the risk for increased work in
tracking too many embargoed reports for long periods of time,
institute a maximum embargo period of 3 months. Also clarify that
the VMT is not responsible for tracking reports of vulnerabilities
in any software maintained outside the OpenStack community, nor
appearing only in extended-maintenance branches, unreleased code in
master branches, pre-releases, release candidates, milestones, and
so on.

Change-Id: Ia501817277a3fed215edd78a8035ad07a90810f0