Turn PowerVMStackers team into a SIG
For the U cycle election there was no candidate for the powervmstackers
team PTL role. This is the second cycle in a row where that problem has
occurred, which indicates that the team has trouble keeping up with the
requirements placed on OpenStack project teams.

A Special Interest Group matches the work of the PowerVMStackers team
better. It places less constraints on the team deliverables, and might
attract more people interested in better support of PowerVM in
OpenStack, which is arguably a special interest.

This patch removes the project team, and moves their code repositories
to be owned by the newly-created PowerVMStackers SIG. That way,
PowerVMStackers remain ATCs and can continue voting for TC election.

Signed-off-by: Doug Hellmann <doug@doughellmann.com>
Signed-off-by: Thierry Carrez <thierry@openstack.org>

Depends-On: https://review.opendev.org/689119
Change-Id: I8cd460ece1b80332386ae41c63970644c4169615