Make TC size configurable
With the goal to reduce the TC to 9, we should be able to configure
the size of the TC in a var.

All the required votes were changed to be counted from the TC size
instead of being staticly defined.

The term "majority" term was appropriately adapted:
sensu stricto it means above 50%. It was replaced by supramajority
(for stronger requirements) or reached enough votes
(for lower requirements).

This made it possible to see the fact that majority was wrongly used
for "formal-vote". The "can_approve = 'YES'" used everywhere
didn't really mean a positive workflow for "formal-vote", as majority
is maybe not reached and votes could be flipped. This separates
a "YES" (a technical 'YES' which could be extended
with the majority condition) from a definitive "CAN APPROVE".

Change-Id: I4682c8795ca8d912ed3fa1951c98a90a34c4352c