Add ansible-collections-openstack to ansible-sig
The ansible community is devolving core modules into "collections"
which can be managed in their own git repos, and do not have to be
located on github.

The ansible-sig has been formed to manage all-things Ansible inside
the OpenStack community so moving the modules into a collection
on OpenDev maintained by the ansible-sig is a good-next-step.

Depends-On: https://review.opendev.org/684738
Change-Id: I0e0e67ef5c76241e78e772f6a8dc8438f5379729
Signed-off-by: Mohammed Naser <mnaser@vexxhost.com>