Adoption of microversion-parse by Oslo team
A recent review of the OpenStack namespace yielded a library that
no project team owns, and yet is specific to openstack and present
in global-requirements: microversion-parse.

After discussing it with cdent, and since it is openstack-specific
and depended on by most openstack projects, it feels like Oslo would
be a good home for it. It would follow an "independent" release model
and I don't think it would trigger much updates or maintenance.

Change-Id: I50f543083b2d1286d4fcf9896b22ef42ad763c04