Begin the 'V' release naming process
We have had several proposals for changing the release
naming process, but have yet to find one that has
agreement that would be better than the current process.

Until we decide on a new naming scheme, we should
continue the process in place. This kicks off the
naming for the V release that will be based on the
collaboration event that will take place in Vancouver
since we will no longer have a Summit event.

Name nominations will use the pattern we have established
and are detailed in:

https://wiki.openstack.org/wiki/Release_Naming/V_Proposals

Change-Id: Ibf3ebb585174aaf5e3e428ce0674f46aa8380dc2
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>