Simplify release naming process
Names can (and should) be proposed by the community, and the TC will then
rank them, and the process can continue as before.

This patch also removes the requirement to tie the name to a geographic
location for events that may or may not happen in line with the cycles.

Change-Id: Ie7d26312628bb7d26bcba39ec24213b3a08b6793
Signed-off-by: Graham Hayes <gr@ham.ie>