Document the chair's prerogative to continue debate
At the PTG we discussed how the chair should handle situations where
there are multiple incompatible patches that each have, on paper,
sufficient votes to be merged. Discussion turned to broader principles
of how we want to operate by consensus, without putting the chair in a
position where they never know if they are authorised to merge a patch
or not. This change to the house rules captures the outcome of that
discussion.

Change-Id: I6187a3ca4ad0438890a8aa88f5228876afd76d43