Make mistral-dashboard a separate deliverable
mistral-dashboard is the only horizon plugin which is not marked as
horizon plugin in the releases repo now as mistral-dashboard is released
as part of the mistral deliverable.

Note that deliverables defined in releases repo should match deliverables
defined in the governance repo and the release team suggested to update
the deliverable definition in the governance repo first (with ACK from
the mistral team).

By making mistral-dashboard a separate deliberable, we can mark
mistral-dashboard as horizon plugin (via the release repo).
Consumers can find mistral-dashboard in the "Horizon Plugins" section
in releases.openstack.org.
In addition, the horizon team uses deliverable files in the release repo
to recognize the list of the horizon plugins so making mistral-dashboard
a separate deliverable would help the horizon team.

A dashboard implementation is usually not tightly coupled with
a corresponding server side implementation, so I believe changing it
to a separate deliverable would have no side effect and the mistral team
can continue the current release model.

Change-Id: I0f93bac85965e6774c500c3580d86ccad0b49f4e