Make validate atcs optional
Without this patch, a change of membership of any ATC could break
the lint jobs.

This is a problem, as the change of membership can happen anytime.

This fixes it by creating a new job to show whether atc validation
fails or not, as non voting. This brings increased visibility to
the eventual membership problems, compared to having them listed
in stdout like we used to do in the past.

Change-Id: I5cb3dee7809d80369eaf8c955695b1389f5eb5d4