Add ideas for OpenStack repository
This is the starting point for the "ideas" project.
"ideas" is a repository where everyone can submit their ideas for
global OpenStack changes.

It is the follow up of a discussion in the last summit in Shanghai,
where we discussed the lack of an easy way to browse and search history
of all the (crazy) ideas for OpenStack floated on the mailing lists.
We established (due to recurring topics) that the MLs are not really
appropriate for this searching and revision control.

Depends-On: https://review.opendev.org/#/c/701530
Change-Id: I8aea45dbd4b52dfedd31965764aa2cf0673f6c3f