Adopt js-openstack-lib into openstacksdk team
The openstack javascript client library has been over in Infra
land for reasons I'm not sure any of us understand. Radosław
Piliszek expressed interest in moving it forward. Since it has
support for sdk concepts such as clouds.yaml, and is a client
library, it seems that it would benefit from maintaining a home
along with like-minded people inside of openstack.

Move the project from Infra to OpenStackSDK.

Change-Id: I794167a5a830398bd1f5c0055aa51ebaa80e5e6e