Begin the 'W' release naming process
This kicks off the naming for the W release. This is the first release
that broadens the scope of release names from being restricted to the
geogrphic region, to allowing any name that begins with the appropriate
letter.

Name nominations will use the pattern we have established and are
detailed in:

https://wiki.openstack.org/wiki/Release_Naming/W_Proposals

Change-Id: I33c561645165dd30866e2f1827c8310d3a46306a
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>